backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
redis[hiredis]==5.2.1
taskiq==0.11.10
taskiq-redis==1.0.2
taskiq-fastapi==0.3.2
aiogram==3.17.0
httpx==0.28.1
sentry-sdk[fastapi]==2.19.2
prometheus-fastapi-instrumentator==7.0.2
pytest==8.3.4
pytest-asyncio==0.24.0
